move types to it own folder
