+++ features/Jobs/components/JobCard.tsx
@@ -2,7 +2,7 @@ import MaterialIcons from '../../../components/MaterialIcons';
 import { imgAlternative } from '../../../utils/assets';
 import { getDifferenceDate } from '../../../utils/dateTime';
 
-import { Job } from '../Job.types';
+import { Job } from '../../../types/Job.types';
 import styles from './JobCard.module.css';
 
 type propTypes = {
